feat: 커플페이지 스켈레톤 코드 작성

+++ src/main/java/dju20/coupleshare/repository/CoupleRepository.java
@@ -3,7 +3,7 @@ package dju20.coupleshare.repository;
 import dju20.coupleshare.entity.Couple;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CoupleRepository extends JpaRepository<Couple, Integer> {
+public interface CoupleRepository extends JpaRepository<Couple, Long> {
 
 }
